// SQCCLIMS/frontend/src/types/rawData.ts
export type RawDataFile = {
  id: number;
  raw_data_id: number;
  original_filename: string;
  stored_filename: string;
  relative_path: string;
  file_path: string;
  file_ext: string;
  mime_type: string;
  file_size: number;
  sha256: string;
  file_role: string;
  preview_supported: number;
  created_at: string;
};